Implement working retrosheet loader
This commit adds a working parser and uploader for Retrosheet play-by-play files. The database table schema had to be changed in some places.

The data is not formatted perfectly -- it had to be manually fixed in some places.

Also, simple-db-migrate has been replaced with YoYo, since simple-db-migrate does not support Python 3.